+++ frontend/desktop/bitwolf/nintendo/gbc_frontend/src/audio.rs
@@ -1,7 +1,7 @@
 use libaudio::AudioContext;
 
 pub struct Audio {
-    ctx: AudioContext<i16, 32>,
+    ctx: AudioContext<f32, 32>,
 }
 
 impl Audio {
@@ -13,7 +13,7 @@ impl Audio {
 }
 
 impl gbc_backend::interfaces::Audio for Audio {
-    fn handle_sample(&mut self, sample: i16) {
+    fn handle_sample(&mut self, sample: f32) {
         self.ctx.push_sample(sample)
     }
 }
